Extract helper functions to separate file
