+++ e2e/Homepage.spec.ts
@@ -1,4 +1,5 @@
-import { test, Page, expect } from "@playwright/test";
+import { test, expect } from "@playwright/test";
+import { getAllProductCards } from "./utils";
 
 test("Wait and get all product cards", async ({ page }) => {
   await page.goto("/");
@@ -62,15 +63,3 @@ test("Testing add to cart functionality", async ({ page }) => {
 
   expect(matchFound).toBe(true);
 });
-
-async function getAllProductCards(page: Page) {
-  await page.waitForSelector('article[test-data="product-card"]', {
-    timeout: 10000,
-  });
-
-  const productCards = page.locator('article[test-data="product-card"]');
-
-  await productCards.first().waitFor({ timeout: 10000 }); // Waits for the first card to be visible
-
-  return productCards;
-}
