+++ e2e/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./productCards";
